third_party/blink/web_tests/http/tests/security/contentTypeOptions/nosniff-script-allowed-expected.txt

CONSOLE MESSAGE: Executed script with MIME type: 'text/javascript1.1'.
CONSOLE MESSAGE: Executed script with MIME type: 'text/javascript1.2'.
CONSOLE MESSAGE: Executed script with MIME type: 'text/javascript1.3'.
CONSOLE MESSAGE: Executed script with MIME type: 'text/livescript'.
CONSOLE MESSAGE: Executed script with MIME type: 'application/ecmascript'.
CONSOLE MESSAGE: Executed script with MIME type: 'application/javascript'.
CONSOLE MESSAGE: Executed script with MIME type: 'application/x-javascript'.
CONSOLE MESSAGE: Executed script with MIME type: 'text/ecmascript'.
CONSOLE MESSAGE: Executed script with MIME type: 'text/javascript'.
CONSOLE MESSAGE: Executed script with MIME type: 'text/jscript'.
Check that script sent with an 'X-Content-Type-Options: nosniff' header is correctly allowed if the MIME type is scripty.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.scriptsSuccessfullyLoaded is scriptyMimeTypes.length
PASS successfullyParsed is true

TEST COMPLETE
